type: make the money-path mypy-clean + Hypothesis invariant tests
Resolves all mypy errors on the money-path (0 remaining under the CI gate).
Two were real latent bugs, not annotation gaps:
  * nn.py  StandardScaler.to_dict() crashed (.tolist() on None) if serialised
    before fit() — now raises a clear ValueError.
  * fx.py  load_fx() could select a NaN missing-ticker sentinel as the index
    source (AttributeError) — tightened to isinstance(pd.Series).
indicators.realized_vol ann widened int->float so the bar-aware annualization
(marks.periods_per_year) type-checks. The rest are Optional/None narrowing.

Adds tests/test_property_invariants.py — Hypothesis property tests asserting the
6 invariants over generated inputs: no-lookahead (perturb the future, past
equity unchanged), cost monotonicity/non-negativity, whole-shares, currency
isolation, P&L reconciliation, sizing bounds, and metric totality.

+++ trading_algo/forex/indicators.py
@@ -48,7 +48,7 @@ def macd(s: pd.Series, fast: int = 12, slow: int = 26, signal: int = 9
     return line, ema(line, signal)
 
 
-def realized_vol(close: pd.Series, window: int, ann: int = 252) -> pd.Series:
+def realized_vol(close: pd.Series, window: int, ann: float = 252) -> pd.Series:
     """Annualised trailing realised volatility from close-to-close returns."""
     rets = close.pct_change(fill_method=None)
     return rets.rolling(window).std() * np.sqrt(ann)
